jobbergate-cli/pyproject.toml
[tool.poetry]
name = "jobbergate-cli"
version = "3.1.0"
description = "Jobbergate CLI Client"
authors = ["Omnivector Solutions <info@omnivector.solutions>"]
license = "MIT"
packages = [ { include = "jobbergate_cli" } ]


[tool.poetry.dependencies]
python = "^3.6.2"
click = "^8.0.1"
inquirer = "^2.7.0"
PyYAML = "^5.4.1"
python-dotenv = "^0.19.0"
sentry-sdk = "^1.4.3"
boto3 = "^1.18.64"
loguru = "^0.6.0"
python-jose = "^3.3.0"
pydantic = "^1.8.2"
yarl = "^1.7.2"
httpx = "^0.22.0"
pendulum = "^2.1.2"
py-buzz = "^3.1.0"
rich = "^11.2.0"
typer = "^0.4.0"
pyperclip = "^1.8.2"
importlib-metadata = "^4.2"
pep562 = "^1.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest-cov = "^3.0.0"
pytest-random-order = "^1.0.4"
pytest-freezegun = "^0.4.2"
pytest-responsemock = "^1.0.1"
isort = "^5.9.3"
pytest-env = "^0.6.2"
respx = "^0.19.2"
plummet = {extras = ["time-machine"], version = "^1.1.0"}
pyproject-flake8 = "^0.0.1-alpha.2"
mypy = "^0.931"
pytest-mock = "^3.6.0"
types-PyYAML = "^6.0.4"
ipython = "^7.10"

[tool.poetry.scripts]
jobbergate = "jobbergate_cli.main:app"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "--random-order",
    "--cov=jobbergate_cli",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
]
env = [
    "ARMADA_API_BASE = https://some-pretend-armada-url.com",
    "JOBBERGATE_DEBUG = false",
    "OIDC_DOMAIN = dummy_auth_domain.com",
    "OIDC_AUDIENCE = https://dummy_auth_audience.com",
    "OIDC_CLIENT_ID = dummy_client_id",
    "OIDC_CLIENT_SECRET = dummy_client_secret",
]

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
lines_after_imports = "2"
profile = "black"
multi_line_output = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 120
max-complexity = 40

[[tool.mypy.overrides]]
module = [
    "urllib3",
    "pyperclip",
    "inquirer.*",
    "jose.*",
    "pep562",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
